requirements.txt
torch
transformers
diffusers
opencv-python
accelerate
pillow
insightface
onnxruntime